refactor: refactor components and layout for improved modularity, simplified data fetching, updated internationalization structure, and enhanced styling consistency across the app.
