+++ src/app/linkbio/_dictionaries/de.json
@@ -1,8 +1,14 @@
 {
   "top-menu": {
-    "language": "Sprache",
-    "share": "Teilen",
-    "copy": "Link kopieren"
+    "menu-label": "Menü",
+    "menu": {
+      "language": "Sprache",
+      "share": "Teilen",
+      "copy": "Link kopieren",
+      "link-text": "Schau dir dieses Linkbio an!"
+    }
   },
-  "toast-feedback": "Link kopiert!"
+  "toast": {
+    "feedback": "Link kopiert!"
+  }
 }
