+++ src/app/linkbio/_dictionaries/pt.json
@@ -1,8 +1,14 @@
 {
   "top-menu": {
-    "language": "Idioma",
-    "share": "Compartilhar",
-    "copy": "Copiar link"
+    "menu-label": "Menu",
+    "menu": {
+      "language": "Idioma",
+      "share": "Compartilhar",
+      "copy": "Copiar link",
+      "link-text": "Confira este Linkbio!"
+    }
   },
-  "toast-feedback": "Link copiado!"
+  "toast": {
+    "feedback": "Link copiado!"
+  }
 }
